Fix lint of day8
Also got lucky and was misunderstood what u8::from(byte) was doing, I
thought it was parsing the u8 number from the ascii byte, but it was not
doing anything. In the end, only relative differences mattered for this
problem, so it worked, but I did fix the parsing logic to actually parse
the number, just to be less "clever".